Fixed bug in Nickname enforcement
Fixed bug in MemberOrchestrator.php to enforce a Discord user name to match to the Seat Main Character name on the role enforcement pass. This is to prevent a discord user from changing their nickname after the initial invite sets their nickname (in the assumption the change nickname discord permission is granted to the user). This is also to automatically update the discord nickname if the SEAT user changes their main character.

The original version would attempt to do this on the role enforcement pass but would only try to do it if a role had changed and then in that case, fail to enforce the nickname as the $options array was never assigned back to by the array_merge() call.