Display Discord Name now updated when Main changed
Fixed an issue where the discord user does not have a nickname set and their username matches the seat assigned nickname generating unneeded update calls. When a nickname is set on a discord user and the nickname is identical to their discord username, it does not set the nick field of the GuildMember block from the discord API. Code will now check both fields.

Added logging for whenever a user actually has a discord update called for tracking when calls are made.