Modified User Policy job retry limit down from 100 to 2
Modified the tries count from 100 down to 2 to reduce spamming the discord API servers. in practice if the call fails once, it will continue failing until discord rate limiting is reset or fail forever. 

This is also to prevent accidentally using up all the Rate limited API calls with an error call which may block out calls that do not cause errors. This commonly happens if the authenticator tries to change something it does not have permission to do resulting in a "403 FORBIDDEN" (happens when it tries to remove a higher ranked discord role or change the nickname of a discord user that has an higher ranked role).

This will also also reduce event stacking as a call that fails every time with the throttling will block the user policy worker threads for 100 [retries] / (20 [allow] * 10 [every X seconds]) = 50 seconds of blocked operation per bad discord user.

This also greatly reduces the log file generated by laravel in the case the call will always fail by not filling it with large numbers of stack trace logging.