Fixed issues in Discord Nickname Enforcement (#1)
* Fixed bug in Nickname enforcement

Fixed bug in MemberOrchestrator.php to enforce a Discord user name to match to the Seat Main Character name on the role enforcement pass. This is to prevent a discord user from changing their nickname after the initial invite sets their nickname (in the assumption the change nickname discord permission is granted to the user). This is also to automatically update the discord nickname and Seat Discord Display name if the SEAT user changes their main character.

The original version would attempt to do this on the role enforcement pass but would only try to do it if a role had changed and then in that case, fail to enforce the nickname as the $options array was never assigned back to by the array_merge() call.

* Fixed issue with discord username matching discord nickname

Fixed an issue where the discord user does not have a nickname set and their username matches the seat assigned nickname generating unneeded update calls. When a nickname is set on a discord user and the nickname is identical to their discord username, it does not set the nick field of the GuildMember block from the discord API. Code will now check both fields.

Added logging for whenever a user actually has a discord update called for tracking when calls are made.

* Modified User Policy job retry limit down from 100 to 2

Modified the tries count from 100 down to 2 to reduce spamming the discord API servers. in practice if the call fails once, it will continue failing until discord rate limiting is reset or fail forever. 

This is also to prevent accidentally using up all the Rate limited API calls with an error call which may block out calls that do not cause errors. This commonly happens if the authenticator tries to change something it does not have permission to do resulting in a "403 FORBIDDEN" (happens when it tries to remove a higher ranked discord role or change the nickname of a discord user that has an higher ranked role).

This will also also reduce event stacking as a call that fails every time with the throttling will block the user policy worker threads for 100 [retries] / (20 [allow] * 10 [every X seconds]) = 50 seconds of blocked operation per bad discord user.

This also greatly reduces the log file generated by laravel in the case the call will always fail by not filling it with large numbers of stack trace logging.

* Code Style Fix

* Don't change nickname unless needed fix

Modified MemberOrchestrator.php to not queue a change to the nickname to discord unless there is actually a change. This reduces the number of changes that need to be made on the discord side and minimizes the effect an edge case where changing the nickname but not the roles would result in a 403 FORBIDDEN error from the Discord API.

* Updated logging to include Discord User Name

Added discord username to logging as the nickname will not be set when the discord username either A) matches their their username exactly or B) is not set.

Added role count to logging the modification of a discord users